requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
python-multipart
openai
chromadb
pypdf
mcp
